flattens interpreter directory
