--- src/interpreter/conditionals/conditionals.rs
@@ -1,4 +1,0 @@
-pub trait Conditionals {
-    fn eval_conditionals(&mut self) -> bool;
-    fn eval_relationships(&self, cond: String) -> bool;
-}
