--- src/interpreter/variables/variables.rs
@@ -1,5 +1,0 @@
-use super::Literal;
-
-pub trait Variables {
-    fn get_var_if_any(&self, var: &Literal) -> Result<Literal, String>;
-}
